refactor: 이름 생성 공통 로직 ObjectPathGenerator로 통일
