+++ api/src/main/kotlin/com/few/api/domain/admin/document/usecase/ConvertContentUseCase.kt
@@ -4,6 +4,7 @@ import com.few.api.domain.admin.document.dto.ConvertContentUseCaseIn
 import com.few.api.domain.admin.document.dto.ConvertContentUseCaseOut
 import com.few.api.domain.admin.document.service.GetUrlService
 import com.few.api.domain.admin.document.service.dto.GetUrlQuery
+import com.few.api.domain.admin.document.utils.ObjectPathGenerator
 
 import com.few.api.repo.dao.document.DocumentDao
 import com.few.api.repo.dao.document.command.InsertDocumentIfoCommand
@@ -12,8 +13,6 @@ import com.few.storage.document.service.PutDocumentService
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 import java.io.File
-import java.time.LocalDate
-import kotlin.random.Random
 
 @Component
 class ConvertContentUseCase(
@@ -24,7 +23,6 @@ class ConvertContentUseCase(
 ) {
     @Transactional
     fun execute(useCaseIn: ConvertContentUseCaseIn): ConvertContentUseCaseOut {
-        val dateDir = LocalDate.now().toString()
         val contentSource = useCaseIn.content
 
         val documentSuffix = contentSource.originalFilename?.substringAfterLast(".") ?: "md"
@@ -33,7 +31,7 @@ class ConvertContentUseCase(
         }.onSuccess {
             contentSource.transferTo(it)
         }.getOrThrow()
-        val documentName = "documents/$dateDir/${generateRandomName()}" + ".$documentSuffix"
+        val documentName = ObjectPathGenerator.documentPath(documentSuffix)
 
         val originDownloadUrl = putDocumentService.execute(documentName, document)?.let { res ->
             val source = res.`object`
@@ -54,16 +52,5 @@ class ConvertContentUseCase(
             convertDocumentService.mdToHtml(document.readBytes().toString(Charsets.UTF_8))
 
         return ConvertContentUseCaseOut(html.replace("\n", "<br>"), originDownloadUrl)
-    }
-
-    private fun generateRandomName(): String {
-        return randomString()
-    }
-
-    private fun randomString(): String {
-        val charPool: List<Char> = ('a'..'z') + ('A'..'Z') + ('0'..'9')
-        return (1..16)
-            .map { Random.nextInt(0, charPool.size).let { charPool[it] } }
-            .joinToString("")
     }
 }
